lxd/instance/operationlock: Remove concept of automatic lock timeout
And adds context to Wait() function.

Rather than require tasks to keep resetting the lock timeout for an operation that takes an
unknown amount of time (which is most of them), which opens the door to unpredicitable behaviour
where another conflicting task can be started whilst the other one is still ongoing.

Instead this takes the view that if a task should timeout after a certain amount of time, it is
up to the task itself to handle its own clean up (including releasing the operation lock).

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>