Document how to enable manual testing of zone DNS.
This CL adds some documentation aiming to reduce friction of learning
how to set up a zone to serve DNS.

Without this information, it is fairly hard for a newcomer: down below
both peers.NAME.address and peers.NAME.key is marked as optional. The
error messages are also unobvious: it is a simple "Transfer failed" on
the client side, leaving it to the imagination of the reader if it failed
because of authorization issues, or because some other configuration is
missing. There is no server side log about it.

Signed-off-by: Baldvin Kovacs <baldvin.kovacs@gmail.com>