lxd/instance/drivers/driver/qemu: Fix addNetDevConfig to match the tap interface settings that qemu uses
When QEMU is passed a tap interface to configure itself it sets unix.IFF_ONE_QUEUE and unix.IFF_VNET_HDR.

LXD was not setting these on the tun interface it was configuring before passing via fd.

This was causing some problems with packet loss and high CPU usage caused by the vhost-net kernel process.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>