lxd/instance/drivers/driver/qemu: Don't load vhost_net module in checkFeatures
The /dev/vhost-net file is available even if the module isn't loaded yet
(if the feature is available in the kernel). The module itself will get
loaded on first use.

So don't explicitly load it as part of feature detection as if VMs are
not used it doesn't need to be loaded.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>