lxd/instance/qemu: Support multiple OVMF firmwares
This defines three types of OVMF firmwares:
 - Generic
 - Secure Boot
 - CSM

For each of those, a set of suitable firmware and variable template
locations are set. The sets are sorted in preference order.

This logic allows LXD to know what variable template and firmware to use
based on configuration (secureboot or csm) as well as to handle a mix of
instances using 2MB or 4MB based firmwares.

Closes #11515

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>