lxd/api: Allow the basic selection of cluster groups when moving instances
When supplying `--target` use the @ character followed by a cluster group name like it is already implemented for the creation of new instances.

Selects the cluster group member based on scheduler hints or falls back to choosing the member with the least amount of instances.

In addition the comments were updated to use the term instances instead of containers.
Alongside this change the `targetMemberInfo` variable is introduced which is also present for the instances creation endpoint and used
to track the target node information during execution if a concrete target was selected or could be resolved from the given group.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>