lxd: Fix deviceEventListener resource scheduling when joining cluster
By passing deviceEventListener a function that can get a fresh copy of state,
rather than using a long held version, which can become out of date.

This manifested itself in no container CPU scheduling being run for instances
launched on a server that had just joined a cluster.

This was because the ServerName property in the state.State object deviceEventListener
was passed on LXD start up contained "none" but after the server had joined the cluster
its ServerName changed to its cluster member name.

This then meant that the instance.LoadNodeAll() call used inside deviceEventListener
was filtering instances that were on server "none", of which there weren't any because
the new instances were associated with the cluster member's name.

This then meant that those instances did not get their CPU scheduling configured until
LXD was next reloaded.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>