lxd/instance/drivers/qemu: Pick a random vsock Context ID
When acquiring a new Context ID for the communication via vsock, use the UUID of the instance as a seed for generating random uint32 candidates. The loop is kept open until a free Context ID is found or the timeout of 5s is reached. The syscall to the vsock returns ENODEV in case the Context ID is not yet assigned.
In case the Context ID of a stopped VM was already acquired again, a new one gets picked.
Removes the `vhost_vsock` feature since the value isn't anymore accessed.

Fixes lxc#11508

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>