lxd/instance/drivers/qemu: Occupy vsock Context ID through syscall
When trying to acquire a new vsock Context ID, we want to occupy it right away to prevent collisions.
This helps for scenarios where two VMs have the same identical volatile.vsock_id (VM export/import) and
in case LXD is used in a nested environment.

If either the acquisition or the syscall fails, the next one gets selected.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>