lxd/storage/backend/lxd: Use the result of FillVolumeConfig in EnsureImage when comparing cached volumes
This allows for logic inside FillVolumeConfig to alter the default settings for new volumes and these then
be used for the comparison with the previous cached image volumes, rather than only considering the pool
config directly.

This is needed for ZFS because zfs.block_mode can be enabled on the pool, but it shouldn't be used for
comparison with VM image volumes because its not relevant for VM image volumes.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>