lxd/firewall: add support for InstanceSetupNetPrio/InstanceClearNetPrio
For xtables we use xt_CLASSIFY target to set skb->priority and
NF_INET_FORWARD hook. Which is not an ideal solution because
it won't work in bridged configuration. But we can do nothing with that.

With nftables we use netdev family and egress hook introduced in
https://github.com/torvalds/linux/commit/42df6e1d221dddc0f2acf2be37e68d553ad65f96
(starting from Linux kernel v5.16)
it allows to catch and process all skb's which go out from the container's veth device.

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>