shared/util/linux: Adds execWrapper for use with ws.MirrorRead() and ws.Mirror()
This wrapper is instantiated using NewExecWrapper() and it wrap an *os.File so
that it uses poll to check if there is something to read first to avoid blocking
reads. It uses a 1s timeout on poll so that if the context is cancelled any
outstanding Read() calls return within 1s.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>