tests: Properly test core.trust_ca_certificates
The test as it was written was quite incorrect.

core.trust_ca_certificates causes valid certificates to be implictly
trusted and not get a certificate store entry.

But as the test was run immediately after a configuration where
core.trust_ca_certificates was not set, there was a leftover entry in
the trust store which the test was then checking.

This corrects the test by clearing the trust store after the initial
test and by further making sure that when using
core.trust_ca_certificates, no trust store entries get created.

Signed-off-by: Stéphane Graber <stgraber@stgraber.org>
Signed-off-by: Mark Laing <mark.laing@canonical.com>