lxd/instance/drivers/driver_qemu: support bootorder in edk2 CSM mode
In the LXD snap we ship a patched version of the CSM edk2 firmware.
Implementation is in the patch ("edk2: force CSM boot mode for bootorder").

Idea behind this is to be able to apply boot priority configuration
for CSM boot mode. While it works perfecly well for UEFI boot
mode, it does not work with CSM mode. It means that CSM boot options
are always come after UEFI boot options, and even worser, manual
changing of the UEFI configuration does not help because bootorder
setting are always reset during reboot of the VM (edk2 maintains
bootorder settings in agreement with what Qemu supplies using
fw_cfg "bootorder" interface).

Usage example:

Suppose that you have a "iso" disk device:
lxc config device add myvm iso disk source=/fullpath/to.iso

And you want to choose it as a first option:
$ lxc config device set myvm iso boot.priority=2

That's all.

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>