lxd/cluster: Retry cluster join if cluster is busy
When lots of nodes are joining the dqlite cluster concurrently, there's
a good chance that some of these nodes will reach out to the leader in
the middle of a role-change. Dqlite reports that a role change has
occcurred with a SQLITE_BUSY error from which we can infer that
we should keep trying to join the cluster until it either succeeds or we
receive some other error.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>