lxd/auth/oidc: Adds methods to set OIDC fields on Verifier.
The rp.RelyingParty is used for authentication via the UI, and the
op.AccessTokenVerifier is used for authentication via the CLI. When
instantiating the rp.RelyingParty, the zitadel library performs
OIDC discovery by calling the /.well-known/openid-configuration endpoint.
To limit the number of calls to this endpoint, we only set a new
rp.RelyingParty if the request host changes (e.g. change of port, DNS
change etc.) as this is required for creating a valid callback URL.

Additionally, we add the "email" scope to the rp.RelyingParty OIDC
scopes.

Signed-off-by: Mark Laing <mark.laing@canonical.com>