test/lint: Add script to invoke golangci-lint with '--new'.
There are thousands of revive rule errors. To avoid having to fix all of
these immediately we can invoke golangci-lint using the '--new' and pass
in a revision. In this script the revision is determined by the most
recent commit in the target branch, defaulting to 'main' for
'GITHUB_BASE_REF' when running in CI.

Signed-off-by: Mark Laing <mark.laing@canonical.com>