doc: decode the objects.inv file
Sphinx creates an objects.inv file that contains a mapping of IDs and
locations in the docs. The UI needs this mapping to be able to resolve
links within the descriptions of config options.

The UI would require additional dependencies to decode the objects.inv
file, so providing the decoded version with the documentation is
easier.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>