lxd/instance/drivers/qemu: Add the `addTargetCpuToMigrationHeader` function
We'd like to export the most accurate view of the target host CPU in order to determine
if the instance running on the source (through host CPU passthrough) is compatible with the
target host. We'll need a way to serialize (through protobuf) the information to the source while doing the
migration negotiation.

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>