test/lint: Temporarily remove `--whole-files` flag.
This flag was added with the intention that we will slowly improve the
quality of the code-base. Some current features are under significant
time pressure and affect a large portion of the code base, and this flag
is quite prohibitive.

It will still not be possible to add new code that causes lint errors.

Signed-off-by: Mark Laing <mark.laing@canonical.com>