client/operations: Don't create events listener after starting an operation
If we implicitly create the events listener after the operation has
fired, then there's a chance the operation will have completed and
emitted the corresponding event before we even begin listening.

Waiting for an operation can thus potentially block indefinitely,
if we first encounter the operation in a running state, but then
it completes before we set up the listener.

Instead, we shouldn't try to create an events listener automatically
when setting up the wait process as we have no way of determining the
current state of an operation yet.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>