lxd: Move handling of OIDC authentication result to new method.
The new method `handleOIDCAuthenticationResult` will check the
identity cache for an identity with the given email address. If
no identiy is found, one is created. If an identity is found but
the name or subject differ from the expected value, the identity
is updated.

If the identities table is modified the cache is updated.

Signed-off-by: Mark Laing <mark.laing@canonical.com>