lxd: Check certificate view permission directly in handler.
Because non-admin users are able to update their own certificate (only
the certificate itself, non of it's LXD properties) permission checks
for certificates tend to happen in the handlers themselves.

In the case of `certificateGet`, writing a separate access handler would
result in an extra database query to expand the fingerprint. So check
permission after getting the full cert to keep things simple.

Signed-off-by: Mark Laing <mark.laing@canonical.com>