lxd/db/cluster: Add missing database entity types.
The `server`, `network_zone`, and `image_alias` entity types were added
in #12872 as they were needed for replacing authorization objects.
However, the database analogues for these entity types were not added
because they are not required for the warning subsystem. This commit
adds these database entity types as they will be needed for
authorization.

Signed-off-by: Mark Laing <mark.laing@canonical.com>