auth/oidc: Ensure the default config expiry interval is set.
When the `oidc.groups.claim` config key was added in #12827, the
logic governing optional configuration of the OIDC verifier was
incorrect. This caused the config expiry interval to be set to zero
if not passed in by the caller.

This caused the verifier to rotate the encryption keys of the relying party
on every call. These keys are used to encrypt the `state` and `pkce`
cookies that OIDC needs to work. As they had been rotated during the
OIDC flow, the callback handler to failed to decrypt the cookies, and so
login was failing.

Signed-off-by: Mark Laing <mark.laing@canonical.com>