lxd/rsync: Merge sendSetup and Send functions
sendSetup can use `RunWrapper` to run rsync from an apparmor profile,
but needs to clean up the profile after the command is done. We track
the running command both in `Send` and `sendSetup` which makes managing
when to begin the cleanup more complex than necessary.

Since this function isn't used anywhere else, and is so tightly coupled
to the Send process, this commit merges the two functions so that we
don't have to track the running command from 2 separate functions before
deleting the profile.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>