api: add a new extension `container_syscall_filtering_allow_deny`
Commit 0c987f66b6f1c added support for the respectful config key names
(`security.syscalls.allow` and `security.syscalls.deny*`) but kept support for
the offensive names for backward compatibility. However, that commit did not
add the `container_syscall_filtering_allow_deny` extension to
`shared/version/api.go`, presumably by mistake.

Now that we are dropping support for the offensively named configuration keys,
use the API extension name and advertise it properly.

Items changed:

* doc/api-extensions: rewrite keys `from container_syscall_filtering` and
  update `container_syscall_filtering_allow_deny` description
* shared/version/api: add `container_syscall_filtering_allow_deny` extension
* lxd/instance/instancetype/instance: remove
  `security.syscalls.(black|white)list*` keys
* lxd/instance/instance_utils: remove `security.syscalls.(black|white)list*`
  keys
* lxd/seccomp/seccomp: remove `security.syscalls.(black|white)list*` keys
  and error out when liblxc doesn't support the needed feature
  `seccomp_allow_deny_syntax`

The change in lxd/seccomp/seccomp requires the underlying liblxc to support the
allowlist/denylist which is the case since
https://github.com/lxc/lxc/commit/78522aa93657841d7384d0800c4acd144a81c9ad
which means lxc 5.0.0 and higher.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>