api: Introduce a new `AgentDeviceRemove` type
Some devices like disk devices with a target path need to be cleanly unmounted.
In the case of a VM instance, we prefer to handle the unmount logic within the agent
and handle the unmounting of the potential overmounts of the guest.

In some cases, we also need to remove a path resource (or not) on the guest side. That is
why we also pass some extra metadata contained in the `Volatile` field, to handle the resource
removal process in the agent and not on the LXD side (this would result in spawning aan SFTP client, which is a wasteful approach).

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>
(cherry picked from commit 3d1e9c7f1b9292858f5779d4932ee4d426f76bfe)
Signed-off-by: Nikita Mezhenskyi <nikita.mezhenskyi@canonical.com>