lxd/identity: Update AuthenticationMethodFromIdentityType method.
api.IdentityTypeOIDCClient was missing from the switch statement. This
meant that calls to IsRestrictedIdentityType and
AuthenticationMethodFromIdentityType would return an error. This error
is ignored in the (*identity.Cache).GetByType method, so it wouldn't be
possible to get OIDC identities from the cache. IsRestrictedIdentityType
will also called from the OpenFGA authorization driver. This was causing
an error every time an OIDC user authenticated.

Signed-off-by: Mark Laing <mark.laing@canonical.com>