lxd/db: Don't propagate expected errors
This was recently refactored here:

https://github.com/canonical/lxd/pull/12754/files#diff-6ba9a4d3cf752f98f07e5a8ff4ba755b19eb813013b0550931532405b46ba9b9

At the end of this block, we return the `err` var.
Previously this block was wrapped in a transaction which was assigned to
an error, but the transaction itself would assign `err` to `nil` if we
received a `sql.ErrNoRows` error. When the transaction was removed, the
`return nil` line was removed as well, and so the `sql.ErrNoRows` error
was being propagated to the caller, instead of `nil`.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>