lxd: Refactor upsertPermissions to use new db methods.
This function was previously quite complicated as we had to check
if a permission already existed before creating it and return a
slice of permission IDs to the caller. Now permissions are directly
related to groups via foreign key we can convert the URLs, delete
any existing permissions for the group, and create the new ones.

Signed-off-by: Mark Laing <mark.laing@canonical.com>