client/lxd/events: Only use events API if client supports auth
The /1.0/events API only allows authenticated connections.

When making a request that spawns an operation, we try to
automatically connect to the events API, but the client may
not necessarily support authentication if it is using a
temporary key to establish trust with a particular endpoint.

This leads to an obscure Forbidden error even though the actual
request we made was trusted. Instead, this ensures we don't try to
incorrectly connect to the events API if the underlying ProtocolLXD
does not support authentication.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>