lxd/project: Update FilterUsedBy function.
The given UsedBy URLs may not contain the "project" query parameter.
Passing them directly to the authorizer is problematic because the
OpenFGA datastore depends on this. This change parses the given URL to
an entity URL and rewrites it, making sure the project parameter is
always set.

Additionally this function is optimised so that if there are multiple
entities of a particular entity type, a permission checker is used
instead of a direct check. This reduces the number of calls to the
authorizer (and therefore the database).

Signed-off-by: Mark Laing <mark.laing@canonical.com>