lxc: Correctly parse remote when listing permissions.
When listing permissions, the first argument (if provided) was always
expected to be a remote. This may not be the case as the remote is
optional and supplementary arguments have the form `<key>=<value>`.

This commit parses the first argument (if given) and uses it as a remote
name if it contains a colon and does not contain an equals sign (used for
filters). Otherwise, the default remote is used.

Signed-off-by: Mark Laing <mark.laing@canonical.com>