lxd: Improves efficiency of operation cancel with permission checker.
Previously, each url was checked with `(Authorizer).CheckPermission`.
With the fine-grained authorization driver this may use more database
calls than is necessary. This commit refactors the permission check to
use a permission checker instead, if there are multiple entities of a
given type. (See similar logic in `lxd/project.FilterUsedBy`).

Signed-off-by: Mark Laing <mark.laing@canonical.com>