lxd/db/cluster: Add identityTypeCertificateMetricsRestricted and identityTypeCertificateMetricsUnrestricted
This effectively replaces identityTypeCertificateMetrics with identityTypeCertificateMetricsRestricted
and adds the new type identityTypeCertificateMetricsUnrestricted.

By replacing identityTypeCertificateMetrics with identityTypeCertificateMetricsRestricted
we ensure that restricted metrics certificates that were falsely moved to identityTypeCertificateMetrics
are still considered to be secure.

On the other side this marks unrestricted certificates as restricted by default.
This has to be manually modified by unrestricting the respective metrics certficiate.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>