doc/server settings: change display of /etc/sysctl.conf settings
While these settings aren't technically config options, I think it can
be beneficial for users to find them in the config option index, and
have a direct link to reference them.

They aren't extracted from the code, so they are included in the
documentation page directly. They will also not be part of the metadata
file.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>