lxd/db: Remove `ErrAlreadyDefined` sentinel error.
This error is imported by the `lxd/request` package so that it can
convert it to an `http.StatusConflict` on a call to
`response.SmartError`. Other applications are using the `lxd/request`
package. Removing this error prevents the import of `lxd/db` into
`lxd/request` and therefore prevents the import of `lxd/db` into
downstream repositories.

Signed-off-by: Mark Laing <mark.laing@canonical.com>