lxd: add forksyscallgo helper
Add "forksyscallgo" helper which is direct analog to
"forksyscall", but written in Golang.

The problem with "forksyscall" is that it intercepts
everything starting from the command line arguments parsing,
which means that everything must be implemented in C.

If you want to call C code from Golang you are good,
but if you want to call Golang from C which intercepts Golang
then you are in trouble.

Signed-off-by: Alexander Mikhalitsyn <aleksandr.mikhalitsyn@canonical.com>