test/suites: It should be possible to grant permissions against a certificate.
This restriction was artificial and also incorrect. We should allow granting
permissions against certificates because this is used by the certificates API.
Since identities and certificates are different entity types, a permission
granted on an identity is not identical to a permission granted on a certificate,
so we need to allow this.

Signed-off-by: Mark Laing <mark.laing@canonical.com>