lxd/db/cluster: Update `EntityType` definition.
`EntityType` is now a wrapper for the new `entityType`
interface. `EntityType` must be a concrete type so that
sql.Scanner and driver.Valuer can be implemented.

`EntityType` itself implements `entityType` by calling
methods on it's `entityType` field. If underlying entity
type is nil, the `EntityType` has the properties of the
"none" entity type, which is used by the warning subsystem.
An `IsNone` method can be used by calling packages to check
if this is the case.

Signed-off-by: Mark Laing <mark.laing@canonical.com>