lxd/db/cluster: Update `EntityType` definition.
`EntityType` is now a wrapper for the new `entityType`
interface. `EntityType` must be a concrete type so that
sql.Scanner and driver.Valuer can be implemented.

`EntityType` itself implements `entityType` by calling
methods on it's `entityType` field. If underlying entity
type is nil, the `EntityType` has the properties of the
"server" entity type, which is used by the warning subsystem.
This means that all warnings that we previously not related to any
entities will now be related to "server" and have a URL of /1.0.

Signed-off-by: Mark Laing <mark.laing@canonical.com>