doc/api extensions: add links to config options
This should cover most config options.

The ones that are still missing are options that exist in different
contexts (for example, for each storage driver). We can't have a
link to those unfortunately.

I've opened https://github.com/canonical/canonical-sphinx-extensions/issues/39
to look into a way of providing a target for such options.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>