lxc/config: Adds a GetInstanceServerWithTransportWrapper method.
Adds a method to get a LXD client that has a transport wrapper
configured. This can be used to e.g. edit the http.Request before it is
sent by the client or inspect the http.Response before any data is
discarded by the client.

Some existing methods have been refactored to avoid duplication.

Signed-off-by: Mark Laing <mark.laing@canonical.com>