Update network_increase_bandwidth.md document with details for u18.04+
The current document does not work properly for Ubuntu 22.04 and probably any version >= 18.04.

I updated this page to include details on how to do this for versions 18.04+ and kept the original section, renaming that to be for versions <= 17.04.

Also added an example command at the bottom on how to change the queue.tx.length in the default LXD profile so that change would apply to all containers.

Signed-off-by: JohnHammell <git@johnhammell.com>