lxd/network/openvswitch/ovn: Workaround OVN load_balancer table index issue by deleting records by UUID
Due to the OVN Load_Balancer tables not having any constraints at the database level, it is possible
for duplicate records with the same name to be added to the database, even though this is not the
default behaviour of lb-add.

If this occurs then lb-del complains that duplicate records exist.

So this commit works around the problem by finding the record(s) by name and then removing by UUID.

Signed-off-by: Thomas Parrott <thomas.parrott@canonical.com>