lxd/storage/drivers/powerflex: Wait for volume to disappear after unmount
This is critical as LXD might already try to mount the volume if two actions
are executed right after another without much delay.
That might cause the volume which is about to be removed to be discovered again.

When requesting the path to a specific volume together with mapVolume set to true,
the cleanup hooks should also contain a step to unmap the volume again from the system.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>