lxc: Unexport all `Run` and `Command` methods (exported).
The "checkPrivateReceivers" argument to the "exported" rule of the
"revive" linter flags all of these methods. This rule was added to
enforce comments e.g. exported methods for interface implementations
where the implementing struct is kept private. The `Run` and `Command`
methods are not called outside of the `lxc` package, so the rule is
correctly indicating that these methods should not be exported.

Signed-off-by: Mark Laing <mark.laing@canonical.com>