lxc: Fix datarace in lxc file get (datarace).
The `io.ReadCloser` returned by the `fileGetWrapper` was captured by
reference. As it was a named return value, this meant that the value of
`buf` could change if the call to `GetInstanceFile` returned after the
user signaled to cancel. This commit removes the named returns so that
the value of `buf` cannot change after the function returns.

Signed-off-by: Mark Laing <mark.laing@canonical.com>