doc: update documentation checks
We should really use the reusable workflows, but this is currently
not possible since we require some non-standard processing.
Therefore, keep the current checks and update them to the latest
framework.

Also, build the docs first and then run the spellcheck on the
docs that are already built (instead of building again).

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>