lxd/network: Use net.IP argument to `checkAddressNotInUse`.
When creating a network forward, the `checkAddressNotInUse` function was
using the value of `listenAddressNet` directly (it is available in
scope), and not the given `net.IP` argument. For automatic IP
allocation, `listenAddressNet` will have been unspecified. This meant
that allocating an IP for a network forward was not correctly validating
the randomly assigned IP addresses and would have led to a bug.

Signed-off-by: Mark Laing <mark.laing@canonical.com>